requirements.txt
streamlit==1.28.1
pdfplumber==0.10.3
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.0.3
numpy==1.24.3
pymongo==4.6.0
python-dotenv==1.0.0
